+++ index.html
@@ -7,5 +7,8 @@
     <link rel="stylesheet" href="style.css" />
     <title>Towerfall</title>
   </head>
-  <body></body>
+  <body>
+    <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.3.0/socket.io.js"></script> -->
+    <!--  <script src="src/index.js"></script> -->
+  </body>
 </html>
