+++ package.json
@@ -15,10 +15,12 @@
     "webpack-merge": "^4.2.2"
   },
   "dependencies": {
+    "express": "^4.17.1",
     "http-server": "^0.12.3",
     "phaser": "^3.54.0",
     "phaser3-weapon-plugin": "^2.1.0",
-    "prettier": "^2.2.1"
+    "prettier": "^2.2.1",
+    "socket.io-client": "^4.1.1"
   },
   "prettier": {
     "singleQuote": true
