--- src/scenes/sceneMain.js
@@ -1,11 +1,0 @@
-/* export class SceneMain extends Phaser.Scene {
-  constructor() {
-    super('SceneMain');
-  }
-  preload() {}
-  create() {
-    console.log('ready!');
-  }
-  update() {}
-}
- */
